// api/src/services/gouv.fr/etablissementType.ts
export interface EtablissementProviderDto {
  identifiant_de_l_etablissement: string;
  nom_etablissement: string;
  type_etablissement: string;
  statut_public_prive: string;
  adresse_1: string;
  adresse_2: any;
  adresse_3: string;
  code_postal: string;
  code_commune: string;
  nom_commune: string;
  code_departement: string;
  code_academie: string;
  code_region: string;
  ecole_maternelle: number;
  ecole_elementaire: number;
  voie_generale: any;
  voie_technologique: any;
  voie_professionnelle: any;
  telephone: string;
  fax: any;
  web: any;
  mail: string;
  restauration: number;
  hebergement: number;
  ulis: number;
  apprentissage: any;
  segpa: any;
  section_arts: any;
  section_cinema: any;
  section_theatre: any;
  section_sport: any;
  section_internationale: any;
  section_europeenne: any;
  lycee_agricole: any;
  lycee_militaire: any;
  lycee_des_metiers: any;
  post_bac: any;
  appartenance_education_prioritaire: any;
  greta: any;
  siren_siret: string;
  nombre_d_eleves: number;
  fiche_onisep: any;
  position: Position;
  type_contrat_prive: string;
  libelle_departement: string;
  libelle_academie: string;
  libelle_region: string;
  coordx_origine: number;
  coordy_origine: number;
  epsg_origine: string;
  nom_circonscription: string;
  latitude: number;
  longitude: number;
  precision_localisation: string;
  date_ouverture: string;
  date_maj_ligne: string;
  etat: string;
  ministere_tutelle: string;
  multi_uai: number;
  rpi_concentre: number;
  rpi_disperse: any;
  code_nature: number;
  libelle_nature: string;
  code_type_contrat_prive: string;
  pial: string;
  etablissement_mere: any;
  type_rattachement_etablissement_mere: any;
  code_circonscription: string;
  code_zone_animation_pedagogique: any;
  libelle_zone_animation_pedagogique: any;
  code_bassin_formation: string;
  libelle_bassin_formation: string;
}

export interface Position {
  lon: number;
  lat: number;
}